requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
future==0.18.3
Jinja2==3.1.2
MarkupSafe==2.1.3
netmiko==4.2.0
ntc-templates==3.5.0
paramiko==3.2.0
pycparser==2.21
PyNaCl==1.5.0
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
PyQt6-WebEngine==6.5.0
PyQt6-WebEngine-Qt6==6.5.1
pyserial==3.5
pywinpty==2.0.10
PyYAML==6.0
qt-material==2.14
scp==0.14.5
six==1.16.0
textfsm==1.1.3
